// src/app/main/models/comercioMovimientosModel.model.ts
export class ComercioMovimientosModel {
    comercioid?: number;
    nombre?: string;
    ruc?: string;
    comerciogrupoid?: number;
    iddepartamento?: string;
    idprovincia?: string;
    iddistrito?: string;
    direccion?: string;
    nombrecontacto?: string;
    telefono?: string;
    correo?: string;
    estado?: string;
}